actor page redraw fix
